create metrics submodule
