+++ docs/source/losses.rst
@@ -6,9 +6,7 @@ kornia.losses
 Reconstruction
 --------------
 
-.. autofunction:: ssim
 .. autofunction:: ssim_loss
-.. autofunction:: psnr
 .. autofunction:: psnr_loss
 .. autofunction:: total_variation
 .. autofunction:: inverse_depth_smoothness_loss
@@ -33,7 +31,6 @@ Module
 .. autoclass:: DiceLoss
 .. autoclass:: TverskyLoss
 .. autoclass:: FocalLoss
-.. autoclass:: SSIM
 .. autoclass:: SSIMLoss
 .. autoclass:: InverseDepthSmoothnessLoss
 .. autoclass:: TotalVariation
