+++ docs/source/metrics.rst
@@ -1,0 +1,29 @@
+kornia.metrics
+==============
+
+.. currentmodule:: kornia.metrics
+
+Module containing metrics for training networks
+
+Classification
+==============
+
+.. autofunction:: accuracy
+
+Segmentation
+============
+
+.. autofunction:: confusion_matrix
+.. autofunction:: mean_iou
+
+Image Quality
+==============
+
+.. autofunction:: psnr
+.. autofunction:: ssim
+.. autoclass:: SSIM
+
+Monitoring
+==========
+
+.. autoclass:: AverageMeter
