+++ kornia/__init__.py
@@ -10,7 +10,9 @@ try:
 except ImportError:
     pass
 
-from kornia import augmentation, color, contrib, enhance, feature, filters, geometry, jit, losses, morphology, utils, x
+from kornia import (
+    augmentation, color, contrib, enhance, feature, filters, geometry, jit, losses, metrics, morphology, utils, x
+)
 from kornia.color import (
     bgr_to_grayscale,
     bgr_to_rgb,
