+++ kornia/geometry/epipolar/__init__.py
@@ -1,6 +1,6 @@
 from .essential import *
 from .fundamental import *
-from .metrics import *
+from ._metrics import *
 from .numeric import *
 from .projection import *
 from .scene import *
