+++ kornia/losses/__init__.py
@@ -3,7 +3,7 @@ from .dice import dice_loss, DiceLoss
 from .divergence import js_div_loss_2d, kl_div_loss_2d
 from .focal import binary_focal_loss_with_logits, BinaryFocalLossWithLogits, focal_loss, FocalLoss
 from .hausdorff import HausdorffERLoss, HausdorffERLoss3D
-from .psnr import psnr, psnr_loss, PSNRLoss
-from .ssim import SSIM, ssim, ssim_loss, SSIMLoss
+from .psnr import psnr_loss, PSNRLoss
+from .ssim import ssim_loss, SSIMLoss
 from .total_variation import total_variation, TotalVariation
 from .tversky import tversky_loss, TverskyLoss
