+++ kornia/metrics/psnr.py
@@ -1,0 +1,49 @@
+import torch
+from torch.nn.functional import mse_loss as mse
+
+
+def psnr(input: torch.Tensor, target: torch.Tensor, max_val: float) -> torch.Tensor:
+    r"""Create a function that calculates the PSNR between 2 images.
+
+    PSNR is Peek Signal to Noise Ratio, which is similar to mean squared error.
+    Given an m x n image, the PSNR is:
+
+    .. math::
+
+        \text{PSNR} = 10 \log_{10} \bigg(\frac{\text{MAX}_I^2}{MSE(I,T)}\bigg)
+
+    where
+
+    .. math::
+
+        \text{MSE}(I,T) = \frac{1}{mn}\sum_{i=0}^{m-1}\sum_{j=0}^{n-1} [I(i,j) - T(i,j)]^2
+
+    and :math:`\text{MAX}_I` is the maximum possible input value
+    (e.g for floating point images :math:`\text{MAX}_I=1`).
+
+    Args:
+        input: the input image with arbitrary shape :math:`(*)`.
+        labels: the labels image with arbitrary shape :math:`(*)`.
+        max_val: The maximum value in the input tensor.
+
+    Return:
+        the computed loss as a scalar.
+
+    Examples:
+        >>> ones = torch.ones(1)
+        >>> psnr(ones, 1.2 * ones, 2.) # 10 * log(4/((1.2-1)**2)) / log(10)
+        tensor(20.0000)
+
+    Reference:
+        https://en.wikipedia.org/wiki/Peak_signal-to-noise_ratio#Definition
+    """
+    if not isinstance(input, torch.Tensor):
+        raise TypeError(f"Expected torch.Tensor but got {type(target)}.")
+
+    if not isinstance(target, torch.Tensor):
+        raise TypeError(f"Expected torch.Tensor but got {type(input)}.")
+
+    if input.shape != target.shape:
+        raise TypeError(f"Expected tensors of equal shapes, but got {input.shape} and {target.shape}")
+
+    return 10.0 * torch.log10(max_val ** 2 / mse(input, target, reduction='mean'))
