--- kornia/utils/metrics/__init__.py
@@ -1,4 +1,0 @@
-from .confusion_matrix import confusion_matrix
-from .mean_iou import mean_iou
-
-__all__ = ["confusion_matrix", "mean_iou"]
