+++ kornia/x/trainer_classification.py
@@ -1,6 +1,6 @@
 import torch
 
-from .metrics import accuracy, AverageMeter
+from kornia.metrics import accuracy, AverageMeter
 from .trainer import Trainer
