+++ test/test_x.py
@@ -5,7 +5,7 @@ from torch.utils.data import Dataset
 
 from kornia.contrib import ClassificationHead, VisionTransformer
 from kornia.x import Configuration, EarlyStopping, ImageClassifierTrainer, ModelCheckpoint
-from kornia.x.metrics import AverageMeter
+from kornia.metrics import AverageMeter
 from kornia.x.utils import TrainerState
